Transport main function in core, sub functuin transfer in envelope
