+++ core/spont_test.go
